refactor: move testspecs package to pkg directory
